requirements.txt
sqlalchemy>=2.0
python-telegram-bot>=20.0
pytest>=8.0
pytest-asyncio>=0.23
ruff>=0.4
